Move region selection to editors mod
